change readdir to Sync & readScenarios to readScenario
