+++ src/app/dataValidator/helpers/dataLoader/readdirSync.test.js
@@ -1,12 +1,12 @@
 global.console.log = jest.fn(); // silence console.log during jest tests
 global.console.time = jest.fn(); // silence console.time during jest tests
 
-const { readdirAsync } = require('./readdirAsync');
+const { readdirSync } = require('./readdirSync');
 
-describe('readdirAsync helper', () => {
-  it('should error if directory passed to readdirAsync does not exist', () => {
+describe('readdirSync helper', () => {
+  it('should error if directory passed to readdirSync does not exist', () => {
     expect(() => {
-      readdirAsync('./././data');
+      readdirSync('./././data');
     }).not.toThrowError();
   });
 });
